Add files via upload
Hello. I have built a program which can detect not only mask but also shows the distance it two persons are detected in the camera. I have used like sklearn for Machine Learning and scipy (scientific-python) for distance estimator. The program will detect mask in run time and if two persons have a distance more the 25 cm( the distance can be change, it was set a/c to laptop's camera) it will display MOVE AWAY!! on the screen.

Press ESC to terminate the program.
It is a jupyter file. Just Run all tabs to see the result.